refactor: move repository and model classes to testutil package for cleaner organization
