+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/service/RsuServiceTest.java
@@ -6,15 +6,15 @@ import com.trihydro.rsuinfobridge.models.tables.RsuModel;
 import com.trihydro.rsuinfobridge.models.tables.RsuOption;
 import com.trihydro.rsuinfobridge.models.tables.SnmpCredential;
 import com.trihydro.rsuinfobridge.models.tables.SnmpProtocol;
-import com.trihydro.rsuinfobridge.repository.RsuCredentialRepository;
-import com.trihydro.rsuinfobridge.repository.RsuIntersectionRepository;
-import com.trihydro.rsuinfobridge.repository.RsuModelRepository;
-import com.trihydro.rsuinfobridge.repository.RsuOptionRepository;
-import com.trihydro.rsuinfobridge.repository.RsuOrganizationRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.RsuCredentialRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.RsuIntersectionRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.RsuModelRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.RsuOptionRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.RsuOrganizationRepository;
 import com.trihydro.rsuinfobridge.repository.RsuRepository;
-import com.trihydro.rsuinfobridge.repository.SnmpCredentialRepository;
-import com.trihydro.rsuinfobridge.repository.SnmpMsgfwdConfigRepository;
-import com.trihydro.rsuinfobridge.repository.SnmpProtocolRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.SnmpCredentialRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.SnmpMsgfwdConfigRepository;
+import com.trihydro.rsuinfobridge.testutil.repository.SnmpProtocolRepository;
 import com.trihydro.rsuinfobridge.testutil.TestcontainersConfiguration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
