+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/model/RsuIntersection.java
@@ -1,4 +1,4 @@
-package com.trihydro.rsuinfobridge.models.tables;
+package com.trihydro.rsuinfobridge.testutil.model;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
