+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/repository/RsuIntersectionRepository.java
@@ -1,9 +1,9 @@
-package com.trihydro.rsuinfobridge.repository;
+package com.trihydro.rsuinfobridge.testutil.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trihydro.rsuinfobridge.models.tables.RsuIntersection;
+import com.trihydro.rsuinfobridge.testutil.model.RsuIntersection;
 
 @Repository
 public interface RsuIntersectionRepository extends JpaRepository<RsuIntersection, Integer> {
