+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/repository/RsuModelRepository.java
@@ -1,4 +1,4 @@
-package com.trihydro.rsuinfobridge.repository;
+package com.trihydro.rsuinfobridge.testutil.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
