+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/repository/RsuOrganizationRepository.java
@@ -1,9 +1,9 @@
-package com.trihydro.rsuinfobridge.repository;
+package com.trihydro.rsuinfobridge.testutil.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trihydro.rsuinfobridge.models.tables.RsuOrganization;
+import com.trihydro.rsuinfobridge.testutil.model.RsuOrganization;
 
 @Repository
 public interface RsuOrganizationRepository extends JpaRepository<RsuOrganization, Integer> {
