+++ services/rsu-info-bridge/src/test/java/com/trihydro/rsuinfobridge/testutil/repository/SnmpMsgfwdConfigRepository.java
@@ -1,10 +1,10 @@
-package com.trihydro.rsuinfobridge.repository;
+package com.trihydro.rsuinfobridge.testutil.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trihydro.rsuinfobridge.models.tables.SnmpMsgfwdConfig;
-import com.trihydro.rsuinfobridge.models.tables.SnmpMsgfwdConfigId;
+import com.trihydro.rsuinfobridge.testutil.model.SnmpMsgfwdConfig;
+import com.trihydro.rsuinfobridge.testutil.model.SnmpMsgfwdConfigId;
 
 @Repository
 public interface SnmpMsgfwdConfigRepository extends JpaRepository<SnmpMsgfwdConfig, SnmpMsgfwdConfigId> {
